6_sae_analysis/requirements.txt
torch
transformers
datasets
peft
unsloth
trl
accelerate
bitsandbytes
vllm
wandb
torchvision
einops
dictionary_learning
ipykernel
matplotlib
pydantic
python-dotenv
